+++ extensions/myst.ui/examples/02_shape/code.js
@@ -10,18 +10,10 @@ myState.createUI = function() {
 	// every created component from here on will default to this context
 	myst.ui.setGlobalContext(this);
 
-	// create some labels
-	this.label1 = new myst.ui.Label({
-		x: 10,
-		y: 10,
-		text: 26,
-		autoResize: true
-	});
-
 	// create some shapes
 	this.shape1 = new myst.ui.Shape({
 		x: 30,
-		y: 60,
+		y: 70,
 		width: 50,
 		height: 50,
 		
@@ -34,7 +26,7 @@ myState.createUI = function() {
 			fill: false,
 			// border thickness. default is 1
 			border: 1,
-			// type of shape. can be any of the following: rectangle, line, triangle, polygon, circle, arc
+			// type of shape. can be any of the following: rectangle, line, triangle, polygon, circle, arc, pie
 			type: 'rectangle',
 			// shape geometry. this is an array of pairs of coordinates in the unit coordinate system (0 to 1)
 			// different shapes may require different number of coordinates. polygon may have any number of coordinates.
@@ -52,7 +44,7 @@ myState.createUI = function() {
 
 	this.shape2 = new myst.ui.Shape({
 		x: 100,
-		y: 60,
+		y: 70,
 		width: 50,
 		height: 50,
 		shape: {
@@ -63,7 +55,7 @@ myState.createUI = function() {
 
 	this.shape3 = new myst.ui.Shape({
 		x: 170,
-		y: 60,
+		y: 70,
 		width: 50,
 		height: 50,
 		shape: {
@@ -76,7 +68,7 @@ myState.createUI = function() {
 
 	this.shape4 = new myst.ui.Shape({
 		x: 240,
-		y: 60,
+		y: 70,
 		width: 50,
 		height: 50,
 		shape: {
@@ -88,22 +80,22 @@ myState.createUI = function() {
 	});
 
 	this.shape5 = new myst.ui.Shape({
-		x: 310,
-		y: 60,
+		x: 312,
+		y: 70,
 		width: 50,
 		height: 50,
 		shape: {
 			type: 'circle',
-			color: 'seagreen',
+			color: 'skyblue',
 			//geometry: [[0.5, 0.5]], // [0.5, 0.5] is the default geometry for circle and arc (center point)
 			fill: false,
-			border: 12
+			border: 2
 		}
 	});
 	
 	this.shape6 = new myst.ui.Shape({
-		x: 30,
-		y: 150,
+		x: 50,
+		y: 170,
 		width: 50,
 		height: 50,
 		shape: {
@@ -116,15 +108,42 @@ myState.createUI = function() {
 	});
 	
 	this.shape7 = new myst.ui.Shape({
-		x: 100,
-		y: 150,
+		x: 120,
+		y: 170,
 		width: 50,
 		height: 50,
 		shape: {
 			type: 'pie',
 			color: 'dodgerblue',
 			parameters: [-90, 180],
+			fill: false,
+			border: 1
+		}
+	});
+	
+	this.shape8 = new myst.ui.Shape({
+		x: 194,
+		y: 160,
+		width: 70,
+		height: 70,
+		shape: {
+			type: 'polygon',
+			color: '#c4b',
+			geometry: [[0.5, 0], [0.17, 1], [1, 0.37], [0, 0.37], [0.83, 1]],
 			fill: false
+		}
+	});
+	
+	this.shape9 = new myst.ui.Shape({
+		x: 280,
+		y: 160,
+		width: 70,
+		height: 70,
+		shape: {
+			type: 'polygon',
+			color: 'orange',
+			geometry: [[0.5, 0], [0.17, 1], [1, 0.37], [0, 0.37], [0.83, 1]],
+			fill: true
 		}
 	});
 
@@ -135,8 +154,8 @@ myState.init = function() {
 };
 myState.draw = function() {
 	this.surface.clear();
-	// draw shapes
 	
+	// draw shapes
 	this.shape1.draw();
 	this.shape2.draw();
 	this.shape3.draw();
@@ -144,11 +163,8 @@ myState.draw = function() {
 	this.shape5.draw();
 	this.shape6.draw();
 	this.shape7.draw();
-
-	//this.paint.rotate(45, [260, 180]);
-	//this.paint.rectFill(150, 150, 100, 100, 'tomato', 150);
-	//this.paint.roundRectangle(150, 150, 100, 100, 'red', 1, 30);
-	//this.paint.restore();
+	this.shape8.draw();
+	this.shape9.draw();
 };
 
 var myGame = new myst.Game({
@@ -159,27 +175,11 @@ var myGame = new myst.Game({
 
 var inputHandler = new myst.Input(myGame);
 
-/*
-var myAssets = {
-	graphics: {
-		button: 'button.png'
-	}
-};
-
-var myLoader = new myst.AssetLoader();
-*/
 
 window.addEventListener('load', function() {
 	// initialize the user interface extension
 	myst.ui.init(inputHandler);
 	// run game
 	myGame.run();
-	/*
-	// load assets
-	myAssets = myLoader.load({
-		assets: myAssets,
-		done: myGame.run // run game when all is loaded
-	});
-	*/
 });