+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+- `FEATURE` Added `sector` and `sectorFill` functions to `Render`.
+- `FEATURE` Added `swap`, `insertionSort`, `heapSort` and `insertSorted` functions.
+
 ## 0.9.4 (February 7, 2021)
 - `BUGFIX` Fixed `Render.circle` and `Render.arc` so they don't continue drawing from old paths.
 - `API CHANGE` Renamed `Render.polygon` to `Render.polygonFill`.